+++ Hospital.Contracts/Dto/Appointment/AppointmentDto.cs
@@ -1,4 +1,4 @@
-﻿namespace Hospital.Contracts.Dto;
+﻿namespace Hospital.Contracts.Dto.Appointment;
 
 /// <summary>
 /// Data transfer object for an appointment.
@@ -9,15 +9,6 @@
 /// <param name="PatientId">The unique id of the patient.</param>
 /// <param name="DoctorId">The unique id of the doctor.</param>
 public record AppointmentDto(
-    DateTime AppointmentTime,
-    string OfficeNumber,
-    bool IsRepeated,
-    Guid PatientId,
-    Guid DoctorId
-);
-
-public record AppointmentResponseDto(
-    Guid Id,
     DateTime AppointmentTime,
     string OfficeNumber,
     bool IsRepeated,
