+++ Hospital.Contracts/Dto/Specialization/SpecializationDto.cs
@@ -1,0 +1,7 @@
+﻿namespace Hospital.Contracts.Dto.Specialization;
+
+/// <summary>
+/// Data transfer object for a specialization.
+/// </summary>
+/// <param name="Name">The name of the specialization.</param>
+public record SpecializationDto(string Name);
