+++ Hospital.Contracts/Dto/Specialization/SpecializationResponseDto.cs
@@ -1,10 +1,9 @@
-﻿namespace Hospital.Contracts.Dto;
+﻿namespace Hospital.Contracts.Dto.Specialization;
 
 /// <summary>
 /// Data transfer object for a specialization.
 /// </summary>
 /// <param name="Name">The name of the specialization.</param>
-public record SpecializationDto(string Name);
 
 public record SpecializationResponseDto(
     Guid Id,
