+++ Hospital.WebApplication/Mappers/DoctorsMapper.cs
@@ -1,5 +1,5 @@
 ﻿using Hospital.Core.Domain.Models;
-using Hospital.Contracts.Dto;
+using Hospital.Contracts.Dto.Doctor;
 
 namespace Hospital.WebApplication.Mappers;
